requirements_backend.txt
setuptools<58
cachetools==5.3.1
numpy==1.24.3
openai==0.27.7
datasets==2.12.0 
sentence-transformers==2.2.2 
python-dotenv==1.0.0
protobuf==3.19.3
pinecone-client[grpc]==2.2.1
google-cloud-firestore==2.7.0
google-cloud-secret-manager==2.12.4
google-cloud-storage==2.9.0
fastapi 
uvicorn[standard]
python-multipart